pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "OVIE"
version = "1.0.0"
description = "Official implementation and models for OVIE (One View Is Enough! Monocular Training for In-the-Wild Novel View Generation)"
requires-python = "==3.10.9"
dependencies = [
    "torch",
    "torchvision",
    "timm",
    "diffusers",
    "accelerate",
    "jupyterlab",
    "notebook",
    "numpy==1.26.1",
    "pillow>=11.3.0",
    "huggingface-hub>=0.34.4",
    "einops>=0.8.1",
    "safetensors>=0.6.2",
    "transformers>=4.57.0.dev0",
    "matplotlib>=3.10.6",
    "scikit-learn>=1.7.2",
    "huggingface_hub",
    "opencv-python",
    "tensorboard>=2.20.0",
    "trimesh>=4.8.2",
    "streamlit>=1.50.0",
    "plotly>=6.3.0",
    "torchviz>=0.0.3",
    "flask>=3.1.2",
    "webdataset==0.2.111",
    "wids>=0.1.11",
    "streamlit-extras>=0.7.8",
    "websockets>=15.0.1",
    "rosbags>=0.11.0",
    "datasets>=4.2.0",
    "tables>=3.10.1",
    "ffmpeg>=1.4",
    "torchcodec==0.7.0",
    "rospkg>=1.6.0",
    "bagpy>=0.4.10",
    "click",
    "opencv-python",
    "scipy",
    "matplotlib",
    "trimesh",
    "pillow",
    "huggingface_hub",
    "numpy",
    "torch>=2.0.0",
    "torchvision",
    "gradio",
    "utils3d @ git+https://github.com/EasternJournalist/utils3d.git@3fab839f0be9931dac7c8488eb0e1600c236e183",
    "pipeline @ git+https://github.com/EasternJournalist/pipeline.git@866f059d2a05cde05e4a52211ec5051fd5f276d6",
    "ftfy>=6.3.1",
    "torchmetrics>=1.8.2",
    "termcolor>=3.1.0",
    "lpips>=0.1.4",
    "pytorch-fid>=0.3.0",
    "wandb>=0.22.3",
    "omegaconf>=2.3.0",
    "moviepy==1.0.3",
    "fairscale>=0.4.13",
]
[tool.setuptools]
py-modules = []

[tool.uv.sources]
transformers = { git = "https://github.com/huggingface/transformers.git" }

[dependency-groups]
dev = [
    "ipykernel>=6.30.1",
    "jupyter>=1.1.1",
]
